frontend/pyproject.toml
[project]
name = "deploy-agentcore-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "strands-agents>=1.19.0",
    "streamlit>=1.52.1",
]
